requirements.txt
pyannote.audio
manim
pandas
pydub
onnxruntime
streamlit
